Cleaning headers
