+++ src/smartpeak/include/SmartPeak/core/SampleGroupProcessor.h
@@ -24,8 +24,6 @@
 #pragma once
 
 #include <SmartPeak/core/Filenames.h>
-#include <SmartPeak/core/MetaDataHandler.h>
-#include <SmartPeak/core/SampleType.h>
 #include <SmartPeak/core/SequenceHandler.h>
 #include <SmartPeak/core/SampleGroupHandler.h>
 #include <SmartPeak/core/Parameters.h>
@@ -65,104 +63,4 @@ namespace SmartPeak
     SampleGroupProcessor() = default;
   };
 
-  struct MergeInjections : SampleGroupProcessor
-  {
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "MERGE_INJECTIONS"; }
-    virtual std::string getDescription() const override { return "Merge multiple injections of the same sample."; }
-    virtual ParameterSet getParameterSchema() const override;
-    virtual std::vector<std::string> getRequirements() const override;
-
-    /**
-      Merge multiple injections of the same sample.
-    */
-    void process(
-      SampleGroupHandler& sampleGroupHandler_IO,
-      const SequenceHandler& sequenceHandler_I,
-      const ParameterSet& params_I,
-      Filenames& filenames_I
-    ) const override;
-
-  private:
-
-    typedef std::tuple<std::string, std::pair<float, float>, float> mergeKeyType;
-    typedef std::pair<std::string, std::string> componentKeyType;
-
-    static bool selectDilutions(
-      const ParameterSet& params,
-      const Filenames& filenames_I,
-      const SequenceHandler& sequenceHandler_I,
-      std::map<componentKeyType, std::map<std::string, std::map<std::set<std::string>, float>>>& component_to_feature_to_injection_to_values);
-
-    static void getMergeKeysToInjections(const SampleGroupHandler& sampleGroupHandler_IO,
-      const SequenceHandler& sequenceHandler_I, 
-      std::set<std::string>& scan_polarities,
-      std::set<std::pair<float, float>>& scan_mass_ranges,
-      std::set<float>& dilution_factors,
-      std::map<mergeKeyType, std::vector<std::set<std::string>>>& merge_keys_to_injection_name);
-
-    static void orderMergeKeysToInjections(std::set<std::string>& scan_polarities,
-      const std::set<std::pair<float, float>>& scan_mass_ranges,
-      const std::set<float>& dilution_factors,
-      std::map<mergeKeyType, std::vector<std::set<std::string>>>& merge_keys_to_injection_name);
-
-    static void getComponentsToFeaturesToInjectionsToValues(const SampleGroupHandler& sampleGroupHandler_IO,
-      const SequenceHandler& sequenceHandler_I,
-      const bool& merge_subordinates,
-      std::map<componentKeyType, std::map<std::string, std::map<std::set<std::string>, float>>>& component_to_feature_to_injection_to_values);
-
-    static void mergeComponentsToFeaturesToInjectionsToValues(const std::string& feature_name, const std::string& merge_rule,
-      const std::set<std::string>& scan_polarities,
-      const std::set<std::pair<float, float>>& scan_mass_ranges,
-      const std::set<float>& dilution_factors,
-      const std::map<mergeKeyType, std::vector<std::set<std::string>>>& merge_keys_to_injection_name,
-      std::map<componentKeyType, std::map<std::string, std::map<std::set<std::string>, float>>>& component_to_feature_to_injection_to_values);
-
-    static void makeFeatureMap(const bool& merge_subordinates,
-      std::set<std::string>& injection_names_set,
-      const std::map<componentKeyType, std::map<std::string, std::map<std::set<std::string>, float>>>& component_to_feature_to_injection_to_values,
-      OpenMS::FeatureMap& feature_map);
-  };
-
-  struct LoadFeaturesSampleGroup : SampleGroupProcessor
-  {
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "LOAD_FEATURES_SAMPLE_GROUP"; }
-    virtual std::string getDescription() const override { return "Load the features for the sample group."; }
-    virtual ParameterSet getParameterSchema() const override;
-
-    /**
-      Load the features for the sample group.
-    */
-    void process(
-      SampleGroupHandler& sampleGroupHandler_IO,
-      const SequenceHandler& sequenceHandler_I,
-      const ParameterSet& params_I,
-      Filenames& filenames_I
-    ) const override;
-
-    /* IFilenamesHandler */
-    virtual void getFilenames(Filenames& filenames) const override;
-  };
-
-  struct StoreFeaturesSampleGroup : SampleGroupProcessor
-  {
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "STORE_FEATURES_SAMPLE_GROUP"; }
-    virtual std::string getDescription() const override { return "Store the features for the sample group."; }
-    virtual ParameterSet getParameterSchema() const override;
-
-    /**
-      Store the features for the sample group.
-    */
-    void process(
-      SampleGroupHandler& sampleGroupHandler_IO,
-      const SequenceHandler& sequenceHandler_I,
-      const ParameterSet& params_I,
-      Filenames& filenames_I
-    ) const override;
-
-    /* IFilenamesHandler */
-    virtual void getFilenames(Filenames& filenames) const override;
-  };
 }
