+++ src/smartpeak/include/SmartPeak/core/SampleGroupProcessors/sources.cmake
@@ -1,0 +1,21 @@
+### the directory name
+set(directory include/SmartPeak/core/SampleGroupProcessors)
+
+### list all header files of the directory here
+set(sources_list_h
+	LoadFeaturesSampleGroup.h
+	MergeInjections.h
+	StoreFeaturesSampleGroup.h
+)
+
+### add path to the filenames
+set(sources_h)
+foreach(i ${sources_list_h})
+	list(APPEND sources_h ${directory}/${i})
+endforeach(i)
+
+### source group definition
+source_group("Header Files\\SmartPeak\\core\\SampleGroupProcessors" FILES ${sources_h})
+
+set(SmartPeak_sources_h ${SmartPeak_sources_h} ${sources_h})
+
